+++ CHANGELOG.md
@@ -4,9 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.29.0] - 2020-08-31
 
+### Added
+- DateTimePicker: Added disabledDates prop. Array of dates and dates ranges that are disabled to be selected as a value.
 
+### Fixed
+- NumericRange: Incorrect 0 value handler.
+- Autocomplete, Input, MultiSelect, Password: hide clear button when disabled.
 
 ## [0.28.0] - 2020-08-17